pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "scan-image-slicer"
authors = [{name = "EMB", email = "eatmyboomstick@gmail.com"}]
readme = "README.md"
license = {file = "LICENSE"}
requires-python = ">=3.8"
keywords = ["scanner", "scan", "slice", "slicer", "image", "photo", "digitize"]
dynamic = ["version", "description"]
dependencies = [
    "pillow",
    "pillow_lut",
    "FreeSimpleGUI",
    "numpy",
    "opencv-python",
    "imutils",
    "ConfigArgParse",
    "tqdm",
    "PyYAML",
    "ruamel.yaml",
    "scipy",
]

classifiers = [
    "Development Status :: 4 - Beta",
    "Operating System :: OS Independent",
    "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "Topic :: Scientific/Engineering :: Image Recognition",
    "Topic :: Scientific/Engineering :: Image Processing",
    "Topic :: Multimedia :: Graphics :: Capture :: Scanners",
]

[project.scripts]
scan-image-slicer = "scan_image_slicer.scan_image_slicer:main"

[project.urls]
Home = "https://github.com/eatmyboomstick/scan_image_slicer"
Source = "https://github.com/eatmyboomstick/scan_image_slicer"